chore: prep colormapregistry for class removal
