+++ DearPyGui/cmake/dpg_sources.cmake
@@ -64,7 +64,6 @@ set(MARVEL_SOURCES
 	"src/ui/AppItems/mvAppItemState.cpp"
 	"src/ui/AppItems/mvAppItem.cpp"
 	"src/ui/AppItems/mvItemRegistry.cpp"
-	"src/ui/AppItems/colors/mvColorMapRegistry.cpp"
 	"src/ui/AppItems/containers/mvCollapsingHeader.cpp"
 	"src/ui/AppItems/containers/mvGroup.cpp"
 	"src/ui/AppItems/containers/mvChild.cpp"
